+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Utxow.hs
@@ -69,7 +69,6 @@ import Data.List.NonEmpty (NonEmpty)
 import Data.Set (Set)
 import Data.Set.NonEmpty (NonEmptySet)
 import GHC.Generics (Generic)
-import NoThunks.Class (InspectHeapNamed (..), NoThunks (..))
 
 -- ================================
 
@@ -174,11 +173,6 @@ deriving instance
   , Eq (PredicateFailure (EraRule "UTXO" era))
   ) =>
   Eq (ConwayUtxowPredFailure era)
-
-deriving via
-  InspectHeapNamed "ConwayUtxowPred" (ConwayUtxowPredFailure era)
-  instance
-    NoThunks (ConwayUtxowPredFailure era)
 
 instance
   ( ConwayEraScript era