+++ libs/cardano-protocol-tpraos/src/Cardano/Protocol/TPraos/Rules/OCert.hs
@@ -24,7 +24,6 @@ import Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
 import Data.Word (Word64)
 import GHC.Generics (Generic)
-import NoThunks.Class (NoThunks (..))
 
 data OCERT c
 
@@ -50,8 +49,6 @@ data OcertPredicateFailure
   | NoCounterForKeyHashOCERT
       (KeyHash BlockIssuer) -- stake pool key hash
   deriving (Show, Eq, Generic)
-
-instance NoThunks OcertPredicateFailure
 
 instance
   ( Crypto c