+++ eras/allegra/impl/CHANGELOG.md
@@ -1,8 +1,9 @@
 # Version history for `cardano-ledger-allegra`
 
-## 1.9.1.0
+## 1.10.0.0
 
 * Add `EraForecast` and `ShelleyEraForecast` instances for `AllegraEra`.
+* Remove `NoThunks` instance for `AllegraUtxoPredFailure`
 
 ## 1.9.0.0